+++ Mailer/extensions/mailer-module/src/Tests/Feature/Api/v1/Handlers/Mailers/MailTypesListingHandlerTest.php
@@ -52,4 +52,31 @@ class MailTypesListingHandlerTest extends BaseApiHandlerTestCase
         $this->assertInstanceOf(\Tomaj\NetteApi\Response\JsonApiResponse::class, $response);
         $this->assertCount(0, $response->getPayload()['data']);
     }
+
+    public function testListWithVariants()
+    {
+        $mailType = $this->createMailTypeWithCategory("category1", "code1", "name1");
+
+        $mailTypeVariant1 = $this->createMailTypeVariant($mailType, 'test1');
+        $mailTypeVariant2 = $this->createMailTypeVariant($mailType, 'test2');
+
+        $params = [];
+        $handler = $this->getHandler(MailTypesListingHandler::class);
+        $response =  $handler->handle($params);
+
+        $this->assertInstanceOf(\Tomaj\NetteApi\Response\JsonApiResponse::class, $response);
+        $this->assertCount(2, $response->getPayload()['data'][0]->variants);
+
+        $this->listVariantsRepository->softDelete($mailTypeVariant1);
+
+        $handler = $this->getHandler(MailTypesListingHandler::class);
+        /** @var \Tomaj\NetteApi\Response\JsonApiResponse $response */
+        $response =  $handler->handle($params);
+
+        $data = $response->getPayload()['data'];
+
+        $this->assertInstanceOf(\Tomaj\NetteApi\Response\JsonApiResponse::class, $response);
+        $this->assertCount(1, $data[0]->variants);
+        $this->assertEquals([$mailTypeVariant2->id => $mailTypeVariant2->title], $data[0]->variants);
+    }
 }